chore: add minimum supported rust version 1.85.0
This change is made to allow cargo to detect what rust versions we want
to support. When `cargo clippy --fix` is run without it, it adds syntax
only supported by the latest rust version 1.88.0.

The 2024 edition requires at least 1.8.5 and since there are no other
issues we can use it as our baseline for now.