style: inline format args (#173)
* chore: add minimum supported rust version 1.85.0

This change is made to allow cargo to detect what rust versions we want
to support. When `cargo clippy --fix` is run without it, it adds syntax
only supported by the latest rust version 1.88.0.

The 2024 edition requires at least 1.8.5 and since there are no other
issues we can use it as our baseline for now.

* style: inline format args

This is a clippy rule that was moved into the style set of lints in
clippy 1.88.0, which is used by default.

Closes: https://github.com/neocmakelsp/neocmakelsp/issues/172